Bring behaviour closer to original game
This commit adds a number of features that
make the behaviour of the game more like
the original arcade game. This includes adding a
hyperspace function which has a chance of
killing the player; removing the FPS counter
and the ability of bullets to kill the player;
making the player larger and slower; adding a
thrust animation to the player; and requiring
the player to repeatedly press the spacebar to
fire.