Improve shooting mechanics
This change significantly slows the player's fire-rate
and somewhat slows the bullets the player fires. This
increases the difficulty in preparation for adding lives.
Collision between shots and asteroids is now detected
using masks rather than rects, which addresses an issue
where shots would destroy an asteroid without appearing
to collide with it.